MNT: Fix horrible bug
I was splitting the full path on the library name.  If any of the
modules or subpackages contain the library name, then the last member
of the split path would simply be the module name.  With this fix,
the full path is split on the library name + the os specific folder
separator.